Thread 분리 위해 boundedElastic 스케쥴러 처리 및 Handler 병합
