+++ src/main/java/team/msa/member/infrastructure/config/WebFluxRouterConfig.java
@@ -8,7 +8,6 @@ import org.springframework.web.reactive.config.WebFluxConfigurer;
 import org.springframework.web.reactive.function.server.RouterFunction;
 import org.springframework.web.reactive.function.server.RouterFunctions;
 import org.springframework.web.reactive.function.server.ServerResponse;
-import team.msa.member.presentation.member.AdminHandler;
 import team.msa.member.presentation.member.MemberHandler;
 
 import static org.springframework.web.reactive.function.server.RequestPredicates.accept;
@@ -17,24 +16,6 @@ import static org.springframework.web.reactive.function.server.RequestPredicates
 @EnableWebFlux // WebFlux 설정 활성화
 public class WebFluxRouterConfig implements WebFluxConfigurer {
 
-    /**
-     * 사이트 운영자 전용 Router
-     * @param adminHandler : 관리자
-     * @return RouterFunction<ServerResponse> : 빌드된 Router
-     */
-    @Bean
-    public RouterFunction<ServerResponse> adminRouterBuilder(AdminHandler adminHandler) {
-
-        return RouterFunctions.route()
-            .path("/admin", adminBuilder ->
-                adminBuilder.nest(accept(MediaType.APPLICATION_JSON), builder ->
-                    builder
-                        .POST("/teacherRegistration", adminHandler::teacherRegistration) // 강사 등록 (관리자만)
-                )
-            )
-            .build();
-    }
-
     @Bean
     public RouterFunction<ServerResponse> memberRouterBuilder(MemberHandler memberHandler) {
 
@@ -42,6 +23,7 @@ public class WebFluxRouterConfig implements WebFluxConfigurer {
             .path("/member", memberBuilder ->
                 memberBuilder.nest(accept(MediaType.APPLICATION_JSON), builder ->
                     builder
+                        .POST("/admin/teacherRegistration", memberHandler::teacherRegistration) // 강사 등록 (관리자만)
                         .POST("/studentRegistration", memberHandler::studentRegistration) // 학생 회원 가입
                         .POST("/login", memberHandler::login) // 로그인
                 )
