+++ src/main/java/team/msa/member/presentation/member/MemberHandler.java
@@ -24,6 +24,22 @@ public class MemberHandler {
     private final MemberApplicationService memberApplicationService;
 
     /**
+     * 강사 등록
+     * --> '사이트 운영자'는 강의 컨테츠를 업로드할 '강사' 회원을 생성할 수 있다.
+     * @param request : 등록할 강사 정보
+     * @return Mono<ServerResponse> : 등록된 강사 정보
+     */
+    public Mono<ServerResponse> teacherRegistration(ServerRequest request) {
+
+        Mono<MemberRegistrationResponse> response = memberApplicationService.memberRegistration(request, MemberType.TEACHER)
+            .subscribeOn(Schedulers.boundedElastic());
+
+        return ok()
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(response, MemberRegistrationResponse.class);
+    }
+
+    /**
      * 학생 회원 가입
      * --> 강의를 수강하고자 하는 사람은 '학생'으로 회원 가입이 가능하다.
      * @param request : 가입할 학생 정보
@@ -31,7 +47,8 @@ public class MemberHandler {
      */
     public Mono<ServerResponse> studentRegistration(ServerRequest request) {
 
-        Mono<MemberRegistrationResponse> response = memberApplicationService.memberRegistration(request, MemberType.STUDENT);
+        Mono<MemberRegistrationResponse> response = memberApplicationService.memberRegistration(request, MemberType.STUDENT)
+            .subscribeOn(Schedulers.boundedElastic());
 
         return ok()
             .contentType(MediaType.APPLICATION_JSON)
